+++ philo_two/includes/philo_two.h
@@ -6,7 +6,7 @@
 /*   By: cdrennan <cdrennan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/06 13:23:40 by cdrennan          #+#    #+#             */
-/*   Updated: 2021/03/15 21:56:29 by cdrennan         ###   ########.fr       */
+/*   Updated: 2021/03/16 00:37:23 by cdrennan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,9 +44,9 @@ typedef struct		s_philo
 typedef struct		s_args
 {
 	int				num_of_philos;
-	int				t_to_die;
-	int				t_to_eat;
-	int				t_to_sleep;
+	long			t_to_die;
+	long			t_to_eat;
+	long			t_to_sleep;
 	int				num_must_eat;
 	long			t_start;
 	sem_t			*forks;
@@ -71,5 +71,5 @@ void				eat(t_philo *philos);
 void				free_all(t_args *args);
 char				*ft_strjoin(char const *s1, char const *s2);
 sem_t				*ft_sem_open(char *name, int value);
-void				ft_usleep(int length);
+void				ft_usleep(long length);
 #endif
